% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_hello_rladies.R
\name{print_hello_rladies}
\alias{print_hello_rladies}
\title{Print hello R-Ladies}
\usage{
print_hello_rladies(x)
}
\arguments{
\item{x}{This argument does nothing.
If it did do something, it might require more words to explain.}
}
\value{
Prints \code{"Hello, R-Ladies"} to the console
}
\description{
This function just prints hello R-Ladies.
}
\examples{
print_hello_rladies(x = 1)

}
